// src/interface/api.ts
import { ZodiacSign } from '@/lib/zodiacSigns';

export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    address: {
        street: string;
        suite: string;
        city: string;
        zipcode: string;
        geo: {
            lat: string;
            lng: string;
        };
    };
    phone: string;
    website: string;
    company: {
        name: string;
        catchPhrase: string;
        bs: string;
    };
}

export interface UserWithZodiac extends User {
    zodiacSign: ZodiacSign;
}

export interface Post {
    userId: number;
    id: number;
    title: string;
    body: string;
}